RfastTOMdist

distance based on the topological overlap map from Ravasz, E., Somera, A., Mongru, D., Oltvai, Z. and Barab´asi, A. (2002). Science Implemented to using the Rfast functions to speed things up since we will be computing this up to 25 times

Usage

RfastTOMdist(A, mat_mult_method = c("Rfast", "RcppEigen"))

Arguments

A

A N x N adjacency matrix where N is the number of genes. values range from -1:1 with higher values indicating highly similar genes. Often correlation ^exponent, but could be angular distance, mutual information or Euclidean distance

mat_mult_method

method for large matrix multiplication, "Rfast" (default) or "RcppEigen" (see details in icwgcna())

Value

A N x N distance matrix with smaller values indicating more related genes.