fastTOMwrapper

Topological overlap map (TOM, from Ravasz, et al (2002). Science) wrapper using the Rfast to speed up calculations. This function can compute TOM for 24K gene matrix in 8 minute of an AWS-EC2 c5.18xlarge instance, though in practice we run it on subsets of most variable genes. Adjacently measure options are Pearson or Spearman correlation raised to a power and angular distance. Note that we only use signed and weighted adjacencies. If users are interested in genes negatively associated with a community they should check community memberships (kME) output from the main function icwgcna().

Usage

fastTOMwrapper(
  X,
  expo = 6,
  Method = c("pearson", "spearman"),
  mat_mult_method = c("Rfast", "RcppEigen")
)

Arguments

X

a gene expression matrix w each column being one sample and N rows representing genes. X should be in log space (usually between 0 and 20)

expo

the power to raise the similarity measure to default = 6. If set to NULL, angular distance is used to applied to the similarity measure ( asin(x) / (pi/2) ).

Method

"pearson" or "spearman" the similarity measure to use

mat_mult_method

method for large matrix multiplication, "Rfast" (default) or "RcppEigen" (see details in icwgcna())

Value

A N x N distance matrix with smaller values indicating more related genes.