Identify Top Gene of Communities that are unique (only belong to one community)

Identify Top Gene of Communities that are unique (only belong to one community)

Usage

find_unique_top_genes(membership_matrix, K = 10, maxIt = 10)

Arguments

membership_matrix

a community membership (kME) matrix with genes as rows and communities as columns. Often community_membership or full_community_membership output from icwgcna()

K

number of unique genes to find.

maxIt

maximum number of iterations to avoid looping to meaningless top genes

Value

data.frame with key = community and value = top gene ID. Top genes will only be associated with one module.

Examples

if (FALSE) { # \dontrun{
library("UCSCXenaTools")
luad <- getTCGAdata(
  project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
  clinical = FALSE, download = TRUE
)
ex <- as.matrix(data.table::fread(luad$destfiles), rownames = 1)
results <- icwgcna(ex)

unique_top_genes <- find_unique_top_genes(
    results$community_membership,
    K = 100,
    maxIt = 10)
} # }