icwgcna

Iterative Correcting Weighted Gene Co-expression Network Analysis function constructing a network from an expression matrix.

Usage

icwgcna(
  ex,
  expo = 6,
  Method = c("pearson", "spearman"),
  q = 0.3,
  maxIt = 10,
  maxComm = 100,
  corCut = 0.8,
  covCut = 0.33,
  mat_mult_method = c("Rfast", "RcppEigen")
)

Arguments

ex

matrix of bulk RNA-seq or microarray gene expression data. This should be in log space and greater than 0.

expo

exponent to use for soft thresholding. If NULL will use angular distance

Method

correlation to use for distance measure, "pearson" (default) or "spearman"

q

quantile (0-1) for first round filtering based on mean expression and standard deviation

maxIt

maximum number of iterations must be 25 or less

maxComm

maximum number of communities to be found

corCut

correlation threshold used for dropping communities

covCut

coefficient of variation (CoV) quantile threshold to use at each iteration for selecting genes to build network. covCut = .667 would use the top third of genes based on CoV after regressing out largest community

mat_mult_method

method for large matrix multiplication, "Rfast" (default) or "RcppEigen" (see details)

Value

Returns a list with the following items:

community_membership - community membership score (kME). Analogous to loadings in PCA.

community_signature - community eigengene, the first principal component the expression of genes in this community. This can be thought of as the average of the scaled expression of top community genes.

full_community_membership - similar to community_membership but includes communities that were dropped in the iterative process.

full_community_signature - similar to community_signature but includes communities that were dropped in the iterative process.

controlled_for - The communities whose signatures were regressed out at each iteration.

Details

Iterative Correcting Weighted Gene Co-expression Network Analysis function for constructing a gene network from a gene expression matrix. The algorithm:

Constructs a signed wgcna network

Drops correlated modules based on kurtosis.

Regresses out the largest community from the expression data.

Repeats steps 1-3 until a maximum number of communities or iterations is reached.

Some differences from standard WGNCA (Horvath/Langfelder)

Makes heavy use of Rfast to compute adjacencies and TOM to enable iterative network creation on > 20K features.

Uses signed adjacency in order to avoid possible distortions of community signatures (eigengenes).

Iteratively regresses out strongest community in order to facilitate discovery of communities possibly obscured larger module(s).

Clustering does not focus on merging communities but dropping to identify strongest module(s).

Enables Spearman correlation for constructing adjacency matrix instead of Pearson to enable robust application in RNA-seq and micro-array data. Future updates may include mutual information

For matrix multiplication the option "Rfast" will use Rfast::mat.mult(), which takes advantage of parallel processing across multiple cores. The option "RcppEigen" will use the RcppEigen engine for C++ code, which tends to be faster when using a single core, but does not take advantage of parallel processing across multiple cores. If running this on a cluster with access to many computer core there is a significant performance advantage to using Rfast::mat.mult()

Examples

if (FALSE) {
library("UCSCXenaTools")
luad <- getTCGAdata(project = "LUAD", mRNASeq = TRUE, mRNASeqType = "normalized",
                  clinical = FALSE, download = TRUE)
ex <- data.table::fread(luad$destfiles)
results <- icwgcna(ex)}